select all main column content as text (not just the first "body")
In the same specimen (pressemitteilung.793657) it was discovered that there can
be reports without an id. The added test addresses the problem of multiple "body"
elements being a possibility but does not offer a solution to cleanly
telling aparat the main (or lack of it) and previous ids (or lack of them).